Post merge fixes - Consolidated both versions of eventUtils.py and utils.py - Fixed imports - Added misc changes skipped during merge - Few documentation additions and renames
