+++ sailbot/config.ini
@@ -1,7 +1,3 @@
-# Stores constants used throughout code
-# There are 2 copies of this file, both should be kept up to date
-# one is located in ./ the other is in ./Main
-# TODO: consolidate
 [MAIN]
 device = pc
 ardu_port = /dev/ttyACM0
@@ -12,14 +8,6 @@ receive_buffer_size = 2048
 baudrate = 115200
 img_accuracy = 10
 log_path = \logs
-
-[MODES]
-MOD_RC = 0
-MOD_COLLISION_AVOID = 1
-MOD_PRECISION_NAVIGATE = 2
-MOD_ENDURANCE = 3
-MOD_STATION_KEEPING = 4
-MOD_SEARCH = 5
 
 [CONSTANTS]
 win_title = Oh God Yep
