+++ sailbot/peripherals/drivers.py
@@ -1,6 +1,8 @@
 """
 hadles turning motors using Odrive/Stepper driver 
 """
+# TO BE REPLACED BY BOATMOVEMENT
+# TODO: REMOVE
 
 import board
 import busio
