+++ sailbot/utils/utils.py
@@ -1,19 +1,15 @@
 """Very commonly used utility functions and classes"""
 import math
-import os
 from dataclasses import dataclass
+from datetime import datetime
+
+import rclpy
+from rclpy.executors import ShutdownException, TimeoutException
+from rclpy.node import Node
 
 from sailbot import constants as c
 from sailbot.peripherals.GPS import GPS
 from sailbot.utils.boatMath import distance_between
-
-DOCKER = os.environ.get("IS_DOCKER", False)
-DOCKER = True if DOCKER == "True" else False
-
-if not DOCKER:
-    pass
-else:
-    pass
 
 
 def singleton(cls):
@@ -54,14 +50,7 @@ class Waypoint:
         self.lon += (dx / EARTH_RADIUS) * (180 / math.pi) / math.cos(self.lat * math.pi / 180)
 
 
-def has_reached_waypoint(waypoint, distance=float(c.config["CONSTANTS"]["reached_waypoint_distance"])):
-    """Returns true/false if the boat is close enough to the waypoint"""
-    a = GPS()
-    boat_gps = Waypoint(a.latitude, a.longitude)
-    return distance_between(boat_gps, waypoint) < distance
-
-
-class dummyObject:
+class DummyObject:
     """
     a class that can be used as a placeholder for something else, usually a physical sensor that is not present.
     allows for setting of arbitrary variables that can be read and written to
@@ -69,3 +58,43 @@ class dummyObject:
 
     def __init__(self, *args, **kwargs):
         pass
+
+
+def has_reached_waypoint(waypoint, distance=float(c.config["CONSTANTS"]["reached_waypoint_distance"])):
+    """Returns true/false if the boat is close enough to the waypoint"""
+    a = GPS()
+    boat_gps = Waypoint(a.latitude, a.longitude)
+    return distance_between(boat_gps, waypoint) < distance
+
+
+def ros_spin_some(node, executor=None, timeout_sec=0, wait_condition=lambda: False):
+    """
+    execute ros callbacks until there are no more available or for timeout_sec, whichever comes first
+    if timeout_sec is 0 then it will execute callbacks until there are no more available
+    """
+    if timeout_sec != 0:
+        endTime = datetime.now() + timeout_sec
+    executor = rclpy.get_global_executor() if executor is None else executor
+    executor.add_node(node)
+    while True:
+        if timeout_sec != 0:
+            remainingTime = endTime - datetime.now()
+            if remainingTime <= 0:
+                break
+        else:
+            remainingTime = 0.0
+
+        try:
+            handler, _, _ = executor.wait_for_ready_callbacks(
+                remainingTime, None, wait_condition
+            )
+        except ShutdownException:
+            pass
+        except TimeoutException:
+            break
+        else:
+            handler()
+            if handler.exception() is not None:
+                raise handler.exception()
+
+    executor.remove_node(node)
