+++ sailbot/virtualPeripherals/camera.py
@@ -10,9 +10,9 @@ import numpy as np
 from rclpy.node import Node
 
 import sailbot.constants as c
-from sailbot.events.eventUtils import Waypoint, distance_between
+from sailbot.utils.boatMath import distance_between
 from sailbot.CV.objectDetection import ObjectDetection, draw_bbox
-from sailbot.utils.utils import singleton
+from sailbot.utils.utils import singleton, Waypoint
 
 
 class Frame:
